pyproject.toml
[build-system]
requires = ["setuptools >= 61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bus-benchmark"
description="A benchmark dataset for bus travel and dwell time prediction"
readme = "README.md"
version = "0.1.0"
requires-python = ">= 3.9.21"
authors = [
    { name = "Alexander Horn", email = "alexander.horn@thi.de" },
    { name = "Philip-Roman Adam", email = "philip-adam@protonmail.com" }
]
keywords = ["public transport", "time series prediction"]
dependencies = [
    'torch == 2.7.0',
    'pandas == 1.5.3',
    'geopandas == 0.13.2',
    'numpy == 1.24.3',
    'scikit-learn == 1.3.2',
    'tqdm == 4.66.1',
    'psycopg2-binary == 2.9.9',
    'lxml == 4.9.3',
    'wandb == 0.17.5',
    'pyarrow == 17.0.0',
    'python-dotenv==1.0.1',
    'tabulate==0.9.0',
    'seaborn==0.13.2',
    'sqlalchemy==1.4.54',
    'ipykernel',
    'matplotlib',
]   

license = { text = "LicenseRef-Proprietary" }

[tool.setuptools.packages.find]
where = ["."]
include = ["bus_benchmark*"]
exclude = ["data", "notebooks", "scripts"]

[tool.black]
safe = true
line-length = 127
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
atomic = true
profile = "black"
src_paths = ["bus_benchmark", "tests", "docs/scripts"]
extra_standard_library = ["typing_extensions"]
indent = 4
lines_after_imports = 2
multi_line_output = 3

[tool.pyright]
include = ["bus_benchmark/**"]
exclude = ["**/__pycache__"]
strict = []

typeCheckingMode = "basic"
pythonVersion = "3.9.21"
pythonPlatform = "All"
enableTypeIgnoreComments = true
